+++ source/dotnet/wholesale-api/Batches/Batches.Application/BatchesClient.cs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Energinet.DataHub.Wholesale.Batches.Application.Model;
 using Energinet.DataHub.Wholesale.Batches.Application.Model.Batches;
 using Energinet.DataHub.Wholesale.Batches.Interfaces;
 using Energinet.DataHub.Wholesale.Batches.Interfaces.Models;
@@ -34,5 +35,58 @@ public class BatchesClient : IBatchesClient
     {
         var batches = await _batchRepository.GetCompletedAfterAsync(completedTime).ConfigureAwait(false);
         return batches.Select(_batchDtoMapper.Map);
+    }
+
+    public async Task<BatchDto> GetAsync(Guid batchId)
+    {
+        var batch = await _batchRepository.GetAsync(batchId).ConfigureAwait(false);
+        return _batchDtoMapper.Map(batch);
+    }
+
+    public async Task<IEnumerable<BatchDto>> SearchAsync(
+        IEnumerable<string> filterByGridAreaCodes,
+        BatchState? filterByExecutionState,
+        DateTimeOffset? minExecutionTime,
+        DateTimeOffset? maxExecutionTime,
+        DateTimeOffset? periodStart,
+        DateTimeOffset? periodEnd)
+    {
+        var executionStateFilter = filterByExecutionState switch
+        {
+            null => Array.Empty<BatchExecutionState>(),
+            BatchState.Pending => new[] { BatchExecutionState.Created, BatchExecutionState.Submitted, BatchExecutionState.Pending },
+            BatchState.Executing => new[] { BatchExecutionState.Executing },
+            BatchState.Completed => new[] { BatchExecutionState.Completed },
+            BatchState.Failed => new[] { BatchExecutionState.Failed },
+            _ => throw new ArgumentOutOfRangeException(nameof(filterByExecutionState)),
+        };
+
+        var gridAreaFilter = filterByGridAreaCodes
+            .Select(g => new GridAreaCode(g))
+            .ToList();
+
+        var minExecutionTimeStart = ConvertToInstant(minExecutionTime);
+        var maxExecutionTimeStart = ConvertToInstant(maxExecutionTime);
+        var periodStartInstant = ConvertToInstant(periodStart);
+        var periodEndInstant = ConvertToInstant(periodEnd);
+
+        var batches = await _batchRepository
+            .SearchAsync(
+                gridAreaFilter,
+                executionStateFilter,
+                minExecutionTimeStart,
+                maxExecutionTimeStart,
+                periodStartInstant,
+                periodEndInstant)
+            .ConfigureAwait(false);
+
+        return batches.Select(_batchDtoMapper.Map);
+    }
+
+    private static Instant? ConvertToInstant(DateTimeOffset? dateTimeOffset)
+    {
+        return dateTimeOffset == null
+            ? null
+            : Instant.FromDateTimeOffset(dateTimeOffset.Value);
     }
 }
