+++ source/dotnet/wholesale-api/WebApi.IntegrationTests/Fixtures/WebApi/WebApiFactory.cs
@@ -64,7 +64,7 @@ public class WebApiFactory : WebApplicationFactory<Startup>
                 ProcessStepResultRepositoryMock?.Object ?? new Mock<ICalculationResultClient>().Object);
             services.AddScoped(_ =>
                 ActorRepositoryMock?.Object ?? new Mock<IActorClient>().Object);
-            services.AddScoped(_ => BatchApplicationServiceMock?.Object ?? new Mock<IGetBatchHandler>().Object);
+            services.AddScoped(_ => BatchesClientMock?.Object ?? new Mock<IBatchesClient>().Object);
 
             var defaultUserContext = new Mock<IUserContext<FrontendUser>>();
             defaultUserContext.Setup(x => x.CurrentUser).Returns(new FrontendUser(Guid.NewGuid(), Guid.NewGuid(), false));
@@ -82,7 +82,7 @@ public class WebApiFactory : WebApplicationFactory<Startup>
 
     public Mock<IActorClient>? ActorRepositoryMock { get; set; }
 
-    public Mock<IGetBatchHandler>? BatchApplicationServiceMock { get; set; }
+    public Mock<IBatchesClient>? BatchesClientMock { get; set; }
 
     public Mock<IUserContext<FrontendUser>>? UserContextMock { get; set; }
