+++ source/dotnet/wholesale-api/WebApi/Configuration/Modules/BatchesRegistration.cs
@@ -62,8 +62,6 @@ public static class BatchesRegistration
         serviceCollection.AddScoped<ICreateBatchHandler, CreateBatchHandler>();
         serviceCollection.AddScoped<IStartCalculationHandler, StartCalculationHandler>();
         serviceCollection.AddScoped<IUpdateBatchExecutionStateHandler, UpdateBatchExecutionStateHandler>();
-        serviceCollection.AddScoped<ISearchBatchHandler, SearchBatchHandler>();
-        serviceCollection.AddScoped<IGetBatchHandler, GetBatchHandler>();
 
         serviceCollection.AddHostedService<StartCalculationWorker>();
         serviceCollection.AddHostedService<UpdateBatchExecutionStateWorker>();
