+++ source/dotnet/wholesale-api/WebApi/V3/ProcessStepResult/ProcessStepResultController.cs
@@ -26,14 +26,14 @@ namespace Energinet.DataHub.Wholesale.WebApi.V3.ProcessStepResult;
 [Route("/v3/batches/{batchId}/processes/{gridAreaCode}/time-series-types/{timeSeriesType}")]
 public class ProcessStepResultController : V3ControllerBase
 {
-    private readonly IGetBatchHandler _batchApplicationService;
+    private readonly IBatchesClient _batchesClient;
     private readonly ICalculationResultClient _calculationResultClient;
 
     public ProcessStepResultController(
-        IGetBatchHandler batchApplicationService,
+        IBatchesClient batchesClient,
         ICalculationResultClient calculationResultClient)
     {
-        _batchApplicationService = batchApplicationService;
+        _batchesClient = batchesClient;
         _calculationResultClient = calculationResultClient;
     }
 
@@ -65,7 +65,7 @@ public class ProcessStepResultController : V3ControllerBase
             energySupplierGln,
             balanceResponsiblePartyGln).ConfigureAwait(false);
 
-        var batch = await _batchApplicationService.GetAsync(batchId).ConfigureAwait(false);
+        var batch = await _batchesClient.GetAsync(batchId).ConfigureAwait(false);
 
         return ProcessStepResultFactory.Create(stepResult, BatchDtoMapper.Map(batch));
     }
